Update to imwheel-0.9.9, which fixes a root exploit in the case where imwheel is setuid root.
Note that the port doesn't install imwheel setuid root, so the
problem would not have affected users of the port / package.

Regardless, it's bad PR to make mouse wheel support so visible
in the FAQ and then keep a known-bad version of the supporting
software in the ports tree.

This update is derived from broken patches received by way of the
maintainer, who no longer has time to maintain the port.  This port is
now looking for a maintainer.

Submitted by:	Jeremy Norris <ishmael27@home.com>