New Port:
This is the port of mplayer, a project which has the effort
to become a very stable and useful movie player.

It is based on the windows codec compilation and
provides the ability to play many formats
in a very good quality without much CPU comsumption.

WWW: http://mplayer.sourceforge.net

PR:		29944
Submitted by:	riggs@f113.hadiko.de