Remove ruby_static from RUN_DEPENDS.
Now that RUBY_ARCH has been changed not to include a minor version of
the system, you no longer need to have ruby_s in order to cope with
the migration problem with a minor system upgrade like 4.3 -> 4.4.

You will need ruby_static only if you are going to do a major upgrade
of your system like 4.3 -> 5.0.