Explicit mention of the patched -release version, and defaulting to the latest one.
Moving HEAD down the menu a bit

Adding additional cvsup servers, and adding state names to the US
servers to help people make a slightly more intelligent choice.

The -P option to cvsup is completely useless with multiplexed mode
and the man page states that it is obsolete.  Cut that.

Changed the comment on scheduling to point to /etc/weekly.local.

Submitted by:	Michael Lucas <mwlucas@blackhelicopters.org>