Bump PORTREVISION for this recent change:
Build Exim with compile-time defaults for exim_user and exim_group.
Use the newly added mailnull user, and the now well-established group
mail.

The use of compile-time defaults for these values is not required for
many sites, but are important for more exotic applications.

This ensures that the packages are useful to the widest audience
possible.