New port: cnslock
Cnslock is a dock application (dockapp) which provides a visual
indication of the states of the three "lock" buttons (caps, num, and
scroll).

It should be especially useful for people who own a wireless keyboard
with no leds on it, and thus have difficulty keeping track of "lock"
states.

Submitted by:	Alexey Dokuchaev