New port: libdict
Libdict is a compact, ANSI C library which provides access to a set of
generic and flexible ``dictionary'' data structures. All algorithms
used in libdict have been optimized, and, with one very small exception,
are not recursive but iterative. It was written by Farooq Mela, and is
released under a BSD style licence.

PR:		50453
Submitted by:	Alex Semenyaka