Add windowlab.
WindowLab is a small and simple window manager, based on aewm, of
novel design.  WindowLab maintains the illusion of direct manipulation
by constraining the mouse pointer when appropriate, ie when a window
cannot be dragged any further in one direction.

It's click-to-focus, shares its window depth policy with the Amiga,
and has a window resizing/reshaping method similar to that of 8 1/2
from Plan 9.

PR:		51168
Submitted by:	 Pierre-Paul Lavoie <ppl@bloodaxis.dyndns.org>