New port: mini_sendmail.
mini_sendmail reads its standard input up to an end-of-file and sends
a copy of the message found there to all of the addresses listed.  The message
is sent by connecting to a local SMTP server.  This means mini_sendmail
can be used to send email from inside a chroot(2) area.