Add tla, an arch CLI written in C.
Arch is a really nifty revision control system. It's "whole-tree
changeset based" which means, roughly, that it can handle (with atomic
commits) file and directory adds, deletes, and renames cleanly, and
that it does branching simply and easily. Arch is also "distributed"
which means, for example that you can make arch branches of your own
from remote projects, even if you don't have write access to the
revision control archives for those projects.

This looks to be as close to an open source p4 replacement as one could
hope without being p4.  I'll go so far as to suggest that if this SCM
was employed by the BSD crowd, merging changes between dragonfly (post
source repo reorog), NetBSD, and OpenBSD would be radically less painful.
It is very possible that the dragonfly fork may not have happened under
the arch SCM development methodology, but if it did, at the very least it
would be possible to incorporate dillion's reorg work in a single patch
set, no cvs admin repo surgery needed.

WWW: http://arch.fifthvision.net/bin/view